requirements.txt
Flask
instagrapi
python-dotenv
gunicorn
requests